pyproject.toml
[project]
name = "fintech-platform"
version = "0.1.0"
description = "Multi-service FastAPI fintech practice stack"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.6.0",
    "pydantic[email]>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "redis>=5.0.0",
    "httpx>=0.27.0",
    "loguru>=0.7.2",
    "prometheus-client>=0.19.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-exporter-otlp>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "uvloop>=0.19.0; sys_platform != 'win32'",
    "tenacity>=8.2.2",
    "pyjwt[crypto]>=2.8.0",
    "python-multipart>=0.0.7",
    "bcrypt>=4.1.2",
    "psycopg[binary]>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.3.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "coverage[toml]>=7.4.0",
    "httpx>=0.27.0",
    "pytest-cov>=5.0.0",
    "faker>=19.0.0",
    "locust>=2.20.0",
    "mypy>=1.8.0",
    "types-redis>=4.6.0.20240218",
    "types-passlib>=1.7.7.20240220",
    "types-python-jose>=3.3.4.20240106",
    "aiosqlite>=0.19.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

# 🧩 Prevent setuptools from auto-discovering random top-level packages
[tool.setuptools]
packages = []
include-package-data = false


# 🧱 uv workspace members
[tool.uv.workspace]
members = [
    "services/api_gateway",
    "services/identity_service",
    "services/wallet_service",
    "services/payments_service",
    "services/risk_service",
    "libs/shared"
]

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I", "N", "UP", "ASYNC", "A", "B", "BLE", "C4", "DTZ", "EM", "RET", "SIM", "T20"]
ignore = ["E203", "E266", "E501"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*" = ["S101"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-q --maxfail=1"
pythonpath = [
    "."
]
